requirements.txt
msal==1.32.3
PyMuPDF==1.26.0
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
